T#7331 - OWL: Make it possible to wrap parts of a library
AML is now wrapped as a separate project, instead of being a part of the AMF wrappers.

Also added a new build method dot_net_library() to openbuild, and a new "dot_net_deps" parameter to the build() method, 
to account for the fact that DLL dependencies for .NET are needed also when compiling, not only when linking.