T#7294 - AAC support in browse AMF part
Implements AAC support in the avformat input/store.

AAC inputs are intended to be used with the packet rate as framerate (375:8 for 48KHz). In order to mix the audio with other sources, 
the frame_rate filter should be used. The frame_rate filter has been updated to provide sample accurate conversion also when seeking.

Since AAC is VBR, an index file is needed to provide seeking. For this purpose, AWI version 4 is introduced, which supports entries 
for different media types such as video and audio. The avformat store will interleave both audio and video entries when creating the 
index. An audio-only avformat input will read the audio entries instead of the video ones.

Updated the FFMpeg version for the osx target.