Modification to the shallow method on the frame_lazy. The new frame_lazy is created from the original frame_lazy, not the inner frame. This is required since the old method broke the following scenario:
Consider a frame_lazy S1, with an inner frame F1. A filter modifies the audio object on S1 when it passes through.
The audio objects on S1 and F1 are now out of sync, but this is not a problem in itself. But, if S1 is shallow copied 
in a pusher the new frame_lazy S2 is created from F1. This means that the audio object in S2 is now not the same as in S1.