BT:25802 - AML WAV plugin reports a few frames shorter duration when reading the input file through the aml: protocol
The WAV plugin relied on the url_seek function to know the amount of header bytes that had been read. However, when using a file reader
that reads data in blocks, this call will give the reading position after the latest read block, which can be further ahead than the
"logical" position. This caused the header size to be overestimated, and consequently the actual WAV data size to be underestimated
and an incorrect duration to be returned.

Fixed by calculating the header size manually instead of using url_seek.