J#APP-298: Handling growing file in VizStenoPlayer
Fixed issues related to loading growing index files. If there is data in the index, but not enough for one entry, the
index parsing would fail, and the index file would be incorrectly treated as an unindexed media file instead.
Also fixed another problem where loading an empty index file would insert a null shared pointer into the index map, 
which would then be returned on all subsequent index requests for that index.