Merged rev. 997 from 1.4 branch:
If there are no output audio tracks in the conform target graph, filter_muxer will call join_peaks() with a lazy_frame that has a null audio object. This caused the frame to be evaluated, degrading performance. Since audio streams are currently not supported anyway, the evaluate() call was removed from get_audio() in frame_lazy.