J#AMF-686: Unable to update Timeline while not playing.
Added a new method to the worker class: cancel_and_remove_job, usable for removing reocurring jobs
which are not waiting to be executed. In this case, just calling set_should_terminate() followed
by wait_for_job_done() will block the entire time while the job is queued, which is probably
not what is intended.

Also bumped minor version to 1.6.1, since this is an interface-breaking change.