J#AMF-914: AMF plays media at very high speed (i.e. no framerate sync) if there is no soundcard is present
The preview store relied on the audio store to provide framerate sync, so if the push to it failed, media
would playback as fast as the machine could handle. One way this can manifest itself is if a USB soundcard
is unplugged.

Fixed by falling back to a manual sleep call if the push to the audio store fails.