J#AMF-1090: Assert in debug mode when creating a 0-channel audio object
The data() and pointer() methods in the audio class dereferenced the
index 0 element of their std::vector in order to get the start address.
This triggers an assert in debug mode for 0-channel objects, since the
vector is empty. Changed these methods so that they now return NULL if
the audio object has 0 channels.

Also removed unnecessary memseti() call in the constructor.