J#AMF-1055: Stenoplayer: Audio drops for a few seconds between chunks
When queried for duration, the avformat input incorrectly used the index to
calculate the result instead of using its internal frame count (which is
kept updated through the sync calls). This led to the input reporting the
full chunk duration to the AMP playlist, even though it couldn't actually
deliver all those frames at the time. The AMP playlist logic then marked
the chunk as complete and did not sync it anymore. This could cause
dropouts of video and/or audio near the end of a chunk.