J#AMF-1000 - Added lowpass filter on rubberband to reduce metallic noise
This commit contains a lowpass filter for riksdagen's stenographers due
to them being distressed about canish noise during slowdown (80% of
normal speed produces this effect on speech). 

The current implementation is a simple walking mean filter with a 
standard backwards look of 14 steps, but that can be adjusted by
the mag=XX flag. One can also write lowpass=0 in order to disable the
lowpass filter for rubberband.

Files changed and why:

- Sconscript                       = Buildscript, added lowpass filter 
                                     to build.
- aml_ardendo.cpp, aml_ardendo.opl = Added the lowpass filter to 
                                     the plugincode.
- plugin_rubberband.cpp            = Added the lowpass filter so as to 
                                     be automatically run.
- filter_lowpass.cpp               = The actuall lowpass implementation.