J#AMF-1133: Transcode AAC encode segfaults on audio from the SVT signal generator with high audio_bit_rate
The last AAC packet out of the MainConcept encoder for the audio in this file was larger than
DEFAULT_PES_PAYLOAD_SIZE in the libavformat ts encoder, which caused a segfault.

Updated to new FFMpeg patched with rev. e31c5ebe1146d98d17a5121312c5444432c81904, which
fixes handling of large audio packets.