J#AMF-1275: Threader filter may hang in rare cases when switching between             playing and paused
wait_for_info() should only exit if the new state is unpaused. If we get
paused while in this method, we should acknowledge it, but not exit.
By exiting, the semantics of pause() were broken, if called while the
background thread was in wait_for_info().
This filter really needs a complete rewrite; the thread interactions are
very unclear and complex to reason about.