J#AMF-1294: Crossfade regression due to property propagation in the compositor filter
This is a regression from the fix for AMF-975. The compositor was changed so that
all the properties were relayed from the overlaid frame to the background. Since
this included properties like x, y, and mix, which controls the composite operation
itself, this led to other compositor filter downstream erroneously applying the same
operation again.

Fixed by not relaying the properties that control the compositor operation.