J#AMF-1554: WAV Audio Conversion Job Stalling on amlbatch
When rewriting the WAV header in the complete() method of the WAV store,
the header parser did not stop when it reached the "data" RIFF block.
Due to this, sample data for the first audio frame were incorrectly
interpreted as header bytes. If the bytes formed a value close to
UINT_MAX, the offset counter would overflow, and parsing continue from an
earlier point in the file which caused this hang.

Fixed by ending the parse when the "data" RIFF block is found.

Also fixed the size value written to the data block, which was 16 bytes
too large.