J#AMF-1737: Putting a filter:decode after a lazy_video/lazy_audio             input and playing backwards corrupts audio
The cause for the issue was the following sequence of events:

1. The lazy input produces an audio block and the internal decode
   filter sets the decoded audio object on the frame.
2. The audio-reversing filter reverses the audio object and sets the
   audio_reversed property on the frame.
3. The frame passes through the decode filter, which now decodes the
   audio block again, replacing the current audio object.
4. The frame now has the reversed property set, but the audio is
   non-reversed, leading to audio corruption when playing.

Fixed by clearing the audio block in the set_audio() method unless the
decoded flag is set, similar to how we clear the stream in
set_image(). Also changed the avdecode filter and the frame_lazy class
so that they set the decoded flag, and the rubberband filter so that
it only does a set_audio() call when playing in reverse (to avoid
clearing the audio_block unnecessarily).