J#AMF-1818: Enhance FTP reading performance in AMF
When loading media files without external AWI indexes through the avformat input, it creates an indexer request for the file. The indexer doesn't know if the uri given points to an AWI or a media file, so it will try to open the media file as a V4, V3 and V2 type AWI before giving up. This slows down opening over all transports, but is particularly expensive with FTP.

* Added a file type hint to the indexer which the inputs supply to it in the request, so that this can be avoided. 
* Changed the avformat input to use the file type hint when requesting an index.