J#AMF-1953: Invalid IMX generated by the avencode filter in AML 1.10
Between the FFMpeg versions used in AML 1.9 and 1.10, the way that the
"non_linear_quant" and "intra_vlc" mpeg2 encoder parameters are specified
changed. Previously, they were set via the "flags2" member on the
AVCodecContext, but they are now generic options instead (set via
av_set_opt_int). The corresponding bit flag constants were removed, but since
the avencode filter picks up the value for "flags2" via the vcodec profiles
instead of setting it directly in code, this wasn't detected.

To fix this, the initialization sequence of the AVCodec and AVCodecContext
objects had to be changed slightly. Previously, the codec context was
initialized with default values, then filled in from the profile, and lastly
used to open the codec. Now, the codec is created when the "video_codec"
profile value is read, and the codec is then used to set the default values on
the codec context. After that, the codec context is filled in from the other
profile values. This was done because the av_set_opt_* methods cannot be used
on a default-initialized context. As a consequence of this, the "video_codec"
value _must_ appear first in the profile, otherwise previously set profile
values will be overwritten.