J#AMF-1961: Compositor filter always triggers re-encode in conform if there are more than one video track
The optimization in the compositor filter which checked if a frame completely
obscured the background, and if so disabled compositing didn't work correctly
if there was more than one incoming video frame. The check needs to loop
through the frames in top-to-bottom order, and if the check succeeds, all the
frames below should be removed (since they aren't visible).

Fixed, and added regression test in the openmedialib test suite.