J#AMF-2123: avformat plugin fails when loading partial ts file with .awi index and packet_stream=1
The size of the .ts file was estimated early on, without using the index. This slightly overestimated
the duration by a few frames. Later on in sync_with_index(), the actual .awi index is used for a
proper sizing. This now produces a shorter duration and failed the sanity check for negative
duration changes.

* Fixed by sizing with the index in initialize as well in order to produce a correct duration.

* Added regression test in the openmedialib test suite under:
  avformat_plugin/amf_2123_failed_to_read_partial_ts.