J#AMF-2124: Regression in growing file support in the avformat plugin             after AMF-1778 when opening files with aml:
When the new ml::io functionality was introduced for J#AMF-1778, the special
functionality that av_read_pause had was missed. When opening files with the
"aml:" protocol (which makes the generic_connector I/O layer in AMF hook into
the file operations), av_read_pause was used as a way to signal to the
generic_connector layer to reopen a file. The new implementation after
J#AMF-1778 did not do this, so there was no file growth anymore when
opening with "aml:".

Since av_read_pause normally has a completely different meaning (to pause
reading of network-based streams), the special functionality is very
confusing and will not be re-introduced. Instead, the generic_connector
layer will now reopen the file on avio_size. This also corresponds to how
the normal avio reader behaves, since it reports file growth on avio_size
without having to be explicitly reopened.