J#AMF-1991: Change the audio class to use the same physical layout for 24-bit audio as it does for 32-bit
Fixed segfault in the aes3 decoder that was mistakenly introduced by the
changes for this ticket. The decoder relies on the destination buffer
having 16 extra padding bytes at the end, but this requirement was very
implicit and "long-distance" (the allocation happened in the openmedialib
audio class).

Reinstated the buffer padding, but inside the aes3 decode function itself
instead.