J#AMF-2262: Calling get_stream on a frame fails silently if the encoder filter   that created the frame has been destroyed
Fixed memory leak due to cyclic reference when the decode filter was used
on a frame that carried both an I-frame stream (i.e. gop size 1) and had
an audio block on it. The frame sent to the audio decoder was not the inner
frame as in frame_lazy::evaluate, but instead the lazy frame itself.
When this frame was cached inside of the pusher input in the audio decoder
subgraph, there was a cyclic reference chain created.

Fixed by pushing the inner frame into the audio decoder subgraph instead.
Also added regression test.